Mark I

+++ VRRocketSuit/shaders/blend.vert
@@ -3,9 +3,9 @@
 // location indices for these attributes correspond to those specified in the
 // InitializeGeometry() function of the main program
 layout(location = 0) in vec3 VertexPosition;
-layout(location = 2) in ivec2 VertexTexCoord;
+layout(location = 2) in vec2 VertexTexCoord;
 
-out ivec2 FragmentTexCoord;
+out vec2 FragmentTexCoord;
 
 void main()
 {
